Move webhooks from apps.py to views.py - can use User model now
