+++ print_nanny_webapp/subscriptions/models.py
@@ -1,6 +1,6 @@
 from django.utils import timezone
 from django.db.models.signals import post_save
-from django.contrib import auth
+from django.contrib.auth import get_user_model
 from django.dispatch import receiver
 from djstripe.enums import SubscriptionStatus
 
@@ -9,7 +9,7 @@ import logging
 
 logger = logging.getLogger(__name__)
 
-User = auth.get_user_model()
+User = get_user_model()
 
 @receiver(post_save, sender=User)
 def start_trial(sender, instance=None, **kwargs):
